capmarket-data-app-main/app/requirements.txt
pandas==2.3.0
numpy==1.26.4
mlflow==3.1.0
langchain==0.3.26
langgraph==0.5.3
databricks-langchain==0.6.0
pydantic==2.11.7
databricks-sdk==0.58.0
databricks-connect==16.1.1
databricks-agents==1.2.0
streamlit==1.46.1